requirements.txt
numpy==2.0.2
pandas==2.2.2
matplotlib==3.10.0
seaborn==0.13.2
opencv-python==4.13.0.90
Pillow==11.3.0
albumentations==2.0.8
scikit-learn==1.6.1
tensorflow==2.19.0

